+++ src/app/cart/cart.module.ts
@@ -1,6 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
@@ -11,8 +9,6 @@ import { SharedModule } from '../shared/shared.module';
     CartItemComponent,
   ],
     imports: [
-        CommonModule,
-        FormsModule,
         SharedModule,
     ],
   exports: [